requirements.txt
celery==5.3.6
Django==4.2.10
django-allauth==0.61.0
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-cron==0.6.0
django-environ==0.11.2
django-import-export==3.3.7
djangorestframework==3.14.0
drf-yasg==1.21.7
pdfkit==1.0.0
pillow==10.2.0
PyJWT==2.8.0
python-crontab==3.0.0
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
weasyprint==60.2
django-filter==23.3
dj-rest-auth==5.0.1
django-celery-results==2.5.1
psycopg2==2.9.9
psycopg-binary==3.1.18
sib-api-v3-sdk==7.6.0
django-ratelimit==4.1.0
channels==4.0.0
daphne==4.1.0
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0
